Use portage_run instead of _domtrans
The advantage of _run is that it will, in the future, include all necessary domains related to portage, not just portage (and this only emerge). One example is layman support. Another advantage is that we can now mention the role as well (in this case system_r) since Portage by default does not run in system_r (it needs to be told through _run).